Je commence à séparer en modules independants.
