+++ src/micro_bench.cpp
@@ -16,6 +16,7 @@
 
 
 #include "utils.h"
+#include "progress.h"
 #include "constants.h"
 #include "traccc_fcts.h"
 #include "bench_mems.hh"
