+++ src/progress.h
@@ -1,0 +1,7 @@
+#pragma once
+
+extern int total_main_seq_runs ;
+extern int current_iteration_count ;
+
+void init_progress() ;
+void print_total_progress()  ;
